+++ dataPumper/src/columnTypes/MultiPolygonColumn.java
@@ -15,11 +15,11 @@ import geometry.Point;
 import geometry.Polygon;
 
 public class MultiPolygonColumn extends IncrementableColumn<MultiPolygon>{
-	private long globalMinX;
-	private long globalMaxX;
+	private double globalMinX;
+	private double globalMaxX;
 	
-	private long globalMinY;
-	private long globalMaxY;
+	private double globalMinY;
+	private double globalMaxY;
 	
 	public MultiPolygonColumn(String name, MySqlDatatypes type, int index) {
 		super(name, type, index);
@@ -28,9 +28,9 @@ public class MultiPolygonColumn extends IncrementableColumn<MultiPolygon>{
 		domainIndex = 0;
 
 		globalMinX = 0;
-		globalMaxX = Long.MAX_VALUE;
+		globalMaxX = Double.MAX_VALUE;
 		globalMinY = 0;
-		globalMaxY = Long.MAX_VALUE;
+		globalMaxY = Double.MAX_VALUE;
 	}
 
 	/**
@@ -62,8 +62,8 @@ public class MultiPolygonColumn extends IncrementableColumn<MultiPolygon>{
 		
 		if( retrievedMultiPolygons.size() != 0 ){
 			
-			globalMinY = globalMinX = Long.MAX_VALUE;
-			globalMaxY = globalMaxX = Long.MIN_VALUE;
+			globalMinY = globalMinX = Double.MAX_VALUE;
+			globalMaxY = globalMaxX = Double.MIN_VALUE;
 			
 			// TODO Fill globals
 			for( MultiPolygon multipolygon : retrievedMultiPolygons ){
@@ -84,12 +84,12 @@ public class MultiPolygonColumn extends IncrementableColumn<MultiPolygon>{
 		}
 		else{
 			min = new MultiPolygon("Multipolygon(((0 0,1 0,1 1,0 1,0 0)))");
-			max = new MultiPolygon("Multipolygon(((0 0,"+Long.MAX_VALUE+" 0,"+Long.MAX_VALUE+" "+Long.MAX_VALUE+",0 "+Long.MAX_VALUE+",0 0)))");
+			max = new MultiPolygon("Multipolygon(((0 0,"+Double.MAX_VALUE+" 0,"+Double.MAX_VALUE+" "+Double.MAX_VALUE+",0 "+Double.MAX_VALUE+",0 0)))");
 			
 			lastInserted = min;
 			
-			globalMaxX = globalMaxY = Long.MAX_VALUE;
-			globalMinX = globalMinY = Long.MIN_VALUE;
+			globalMaxX = globalMaxY = Double.MAX_VALUE;
+			globalMinX = globalMinY = Double.MIN_VALUE;
 		}
 		
 	}
